requirements.txt

# 深度学习框架
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0

# Transformer和多语言模型
transformers>=4.30.0
tokenizers>=0.13.3

# EEG数据处理
mne>=1.4.0
pywt>=1.4.1

# 科学计算
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# 数据处理
pandas>=2.0.0
h5py>=3.8.0

# 评估指标
nltk>=3.8.1
rouge-score>=0.1.2

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 实验记录
wandb>=0.15.0
tensorboard>=2.13.0

# 配置管理
pyyaml>=6.0
omegaconf>=2.3.0

# 进度条和工具
tqdm>=4.65.0
colorama>=0.4.6

# 开发工具
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# 其他依赖
Pillow>=9.5.0
requests>=2.31.0